pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deploy-kit"
version = "0.1.0"
description = "Opinionated GCP deploy CLI for Cloud Run, Cloud Run Jobs, and Firebase Hosting"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1.0",
  "python-dotenv>=1.0.0",
  "google-api-python-client>=2.130.0",
  "google-cloud-core>=2.3.3",
  "google-cloud-artifact-registry>=1.11.3",
  "google-cloud-secret-manager>=2.16.4",
  "google-cloud-run>=0.10.4",
  "google-cloud-storage>=2.17.0",
  "google-cloud-bigquery>=3.25.0",
  "google-cloud-sqlcommenter>=2.0.0; python_version<'3.13'",  # optional, safe upper bound
]

[project.scripts]
deploy-main = "deploy_kit.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["deploy_kit*"]


